+++ getProductsList/src/functions/getProductsList/index.ts
@@ -6,7 +6,7 @@ export default {
     {
       http: {
         method: 'get',
-        path: 'products',
+        path: 'products/',
         cors: true,
       },
     },
