--- getProductsList/src/functions/getProductsListFromRDS/index.ts
@@ -1,14 +1,0 @@
-import { handlerPath } from '@libs/handlerResolver';
-
-export default {
-  handler: `${handlerPath(__dirname)}/handler.main`,
-  events: [
-    {
-      http: {
-        method: 'get',
-        path: 'products/sql',
-        cors: true,
-      },
-    },
-  ],
-};
