+++ getProductsList/src/functions/index.ts
@@ -1,5 +1,3 @@
 export { default as getProductsList } from './getProductsList';
 export { default as getProductById } from './getProductById';
-export { default as getProductsListFromRDS } from './getProductsListFromRDS'
-export { default as getProductByIdFromRDS } from './getProductByIdFromRDS'
-export { default as postProductToRDS } from './postProduct'
+export { default as postProduct } from './postProduct'
